added deletion method  real time

+++ react_chat_app/src/appwrite/backgroundDelete.js
@@ -3,13 +3,18 @@ import service from "./database";
 const bgDelete = async()=>
 {
     let res = await service.getMessages();
-    let l = 5;
+    let l = res.documents.length -  25;
+
     console.log("before "  , res.documents)
-    for (let i = 0 ; i < l  ; i++){
-        await service.deleteMessage(res.documents[i].$id);
-        console.log("deleted" , res.documents[i].$id)
+
+    if(l > 0){
+        for (let i = 0 ; i < l  ; i++){
+            await service.deleteMessage(res.documents[i].$id);
+            console.log("deleted" , res.documents[i].$id)
+        }
+        console.log(res.documents)
     }
-    console.log(res.documents)
+    
 }
 
 export default bgDelete;
